backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
pytz==2024.1
gunicorn==22.0.0
dj-database-url==2.1.0
whitenoise==6.6.0
